# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Listar proyectos",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (active, inactive)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre parcial",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from_date",
                        "in": "query",
                        "description": "Fecha inicial de creación (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to_date",
                        "in": "query",
                        "description": "Fecha final de creación (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de proyectos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Crear nuevo proyecto",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto Nuevo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción opcional"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Mostrar detalle de proyecto",
                "operationId": "865085b82af27d647cf1f2a772aa7927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del proyecto"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Actualizar proyecto",
                "operationId": "fe600b7ffce78456bce67c51b9eca51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Proyecto Actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "inactive"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Proyectos"
                ],
                "summary": "Eliminar proyecto",
                "operationId": "d81ac2b0b51fa7143c69f1a5f4c27eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del proyecto (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proyecto eliminado"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Listar tareas con filtros",
                "operationId": "be68905d866e856db6c738172b8d929a",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Estado de la tarea",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Prioridad de la tarea",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "due_date",
                        "in": "query",
                        "description": "Fecha límite",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "ID del proyecto",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de tareas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "ebbaabc58e3db61374a57c53544959b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Tarea A"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción opcional"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener detalle de una tarea",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la tarea"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Actualizar una tarea",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title",
                                    "status",
                                    "priority",
                                    "due_date"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "priority": {
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Eliminar una tarea",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada exitosamente"
                    },
                    "404": {
                        "description": "Tarea no encontrada"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Proyectos",
            "description": "API RESTFul de Proyectos"
        },
        {
            "name": "Tareas",
            "description": "Operaciones CRUD para tareas"
        }
    ]
}